+++ src/Aks/Aks/Commands/NewOrUpdateAgentPoolBase.cs
@@ -12,7 +12,11 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Microsoft.Azure.Commands.Aks.Commands;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
+using Microsoft.Azure.Management.ContainerService;
+using Microsoft.Azure.Management.ContainerService.Models;
+using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using Newtonsoft.Json;
 using System.Collections;
 using System.Collections.Generic;
@@ -50,5 +54,21 @@ namespace Microsoft.Azure.Commands.Aks
 
         [Parameter(Mandatory = false, HelpMessage = "The node taints added to new nodes during node pool create and scale")]
         public string[] NodeTaint { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "Aks custom headers")]
+        public Hashtable AksCustomHeader { get; set; }
+
+        private protected AgentPool CreateOrUpdate(string resourceGroupName, string resourceName, string agentPoolName, AgentPool parameters)
+        {
+            if (this.IsParameterBound(c => c.AksCustomHeader))
+            {
+                Dictionary<string, List<string>> customHeaders = Utilities.HashtableToDictionary(AksCustomHeader);
+                return Client.AgentPools.CreateOrUpdateWithHttpMessagesAsync(resourceGroupName, resourceName, agentPoolName, parameters, customHeaders).GetAwaiter().GetResult().Body;
+            }
+            else
+            {
+                return Client.AgentPools.CreateOrUpdate(resourceGroupName, resourceName, agentPoolName, parameters);
+            }
+        }
     }
 }