+++ src/Aks/Aks/ChangeLog.md
@@ -24,6 +24,8 @@
 * Added property `-NodeImageVersion` for the output of `Get-AzAksNodePool`[#19893]
 * Added parameter `-NodePoolLabel` for `Set-AzAksCluster`, `-NodeLabel` for `New-AzAksNodePool` and `Update-AzAksNodePool`
 * Added parameter `-NodePoolTag` for `New-AzAksCluster` and `Set-AzAksCluster`, `-Tag` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added parameter `-NodeTaint` for `New-AzAksNodePool` and `Update-AzAksNodePool`
+* Added parameter `-AksCustomHeader` for `Set-AzAksCluster`, `New-AzAksNodePool` and `Update-AzAksNodePool`
 
 ## Version 5.0.1
 * Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]